requirements.txt
fastapi
uvicorn
gradio
httpx
requests
python-multipart
Pillow
numpy
opencv-python
google-genai
torch
sqlalchemy
moviepy